Correct updatefile.txt version mismatch
The notes in updatefile.txt are correct, however the version number is not. Thus the updater downgrades the version of the plugin.
It also seems that 4.4.5 which changes up the debug command is actually 4.4.4 in the .sp file, while the plugin is 4.4.5 in the .sp file of the 4.4.6 release. This is also corrected.